get stack run config based on template name
